requirements.txt
Django>=4.2.6,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<2.10
Pillow>=10.0.1,<10.1